+++ tests/rollup/index.js
@@ -1,0 +1,7 @@
+import Template from './template.hbs'
+
+function load() {
+    let output = Template()
+    return output
+}
+load()