+++ CHANGELOG.md
@@ -128,12 +128,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [1.1.3]  2020-8-29
+## [1.1.3](https://github.com/JohannIsaac/rollup-plugin-handlebars-compiler/pull/16)  2020-8-29
 
 ### Added
 
+- Added support for nested partials
 - Unit tests for referencing partials from ancestor/parent/cousin-dirs
 
 ### Changed
 
+- Fix: Fixed critical bug where passing no plugin options throws an error
 - Fix: Removed registering the context template file as a partial